IGAPP-985: Fix test
